#27 Added validation to PartyService

+++ src/main/java/team/asd/service/PartyService.java
@@ -1,7 +1,9 @@
 package team.asd.service;
 
+import org.apache.commons.lang3.StringUtils;
 import team.asd.dao.PartyDao;
 import team.asd.entity.Party;
+import team.asd.exceptions.ValidationException;
 
 public class PartyService {
 	private final PartyDao partyDao;
@@ -10,19 +12,31 @@ public class PartyService {
 		this.partyDao = partyDao;
 	}
 
-	public Party readById(Integer id) {
+	public Party readById(Integer id) throws ValidationException {
+		if (id == null || id <= 0) {
+			throw new ValidationException("Wrong id was provided");
+		}
 		return partyDao.readById(id);
 	}
 
-	public void create(Party party) {
+	public void create(Party party) throws ValidationException {
+		if (party == null || StringUtils.isBlank(party.getName())) {
+			throw new ValidationException("Wrong Party object was provided");
+		}
 		partyDao.create(party);
 	}
 
-	public void update(Party party) {
+	public void update(Party party) throws ValidationException {
+		if (party == null || party.getId() == null || party.getId() <= 0) {
+			throw new ValidationException("Wrong Party object was provided");
+		}
 		partyDao.update(party);
 	}
 
-	public void delete(Integer id) {
+	public void delete(Integer id) throws ValidationException {
+		if (id == null || id <= 0) {
+			throw new ValidationException("Wrong id was provided");
+		}
 		partyDao.deleteById(id);
 	}
 }
